requirements.txt
bottle
gunicorn
PyYAML
pyopenssl
m2crypto
coloredlogs
pyocclient
